pyproject.toml
[build-system]
requires = [
    "setuptools",
]
build-backend = "setuptools.build_meta"

[project]
requires-python = ">=3.12"
name = "sf_kedro"
readme = "README.md"
dynamic = [
    "version",
]
dependencies = [
    # SignalFlow ecosystem
    "signalflow-trading>=0.5.0",
    "signalflow-ta>=0.5.0",
    "signalflow-nn>=0.5.0",
    # Kedro
    "kedro~=1.1.1",
    "kedro-datasets[pickle,text]>=5.0",
    # MLOps
    "mlflow>=2.0",
    "optuna>=3.0",
    # Jupyter
    "ipython>=8.10",
    "jupyterlab>=4.0",
    "notebook",
]

[project.scripts]
sf-kedro = "sf_kedro.__main__:main"

[project.entry-points."kedro.hooks"]

[project.optional-dependencies]
dev = [
    "pytest-cov>=3,<7",
    "pytest-mock>=1.7.1, <2.0",
    "pytest~=7.2",
    "ruff~=0.12.0",
]

[tool.setuptools.dynamic.version]
attr = "sf_kedro.__version__"

[tool.setuptools.packages.find]
where = [
    "src",
]
namespaces = false

[tool.kedro]
package_name = "sf_kedro"
project_name = "SF Kedro"
kedro_init_version = "1.1.1"
tools = "['Testing', 'Custom Logging', 'Data Structure']"
example_pipeline = "False"
source_dir = "src"

[tool.pytest.ini_options]
addopts = "--cov-report term-missing --cov src/sf_kedro -ra"

[tool.coverage.report]
fail_under = 0
show_missing = true
exclude_lines = [
    "pragma: no cover",
    "raise NotImplementedError",
]

[tool.ruff]
target-version = "py312"
line-length = 120
src = ["src"]
extend-exclude = ["notebooks"]

[tool.ruff.lint]
select = ["E", "W", "F", "I", "UP", "B", "SIM", "RUF"]

[tool.ruff.lint.per-file-ignores]
"scripts/*" = ["E402"]

[tool.ruff.lint.isort]
known-first-party = ["sf_kedro"]

[tool.kedro_telemetry]
project_id = "455c4b21f6d7499ba15d922f41f386fa"
